requirements.txt
numpy==2.4.2
pandas==3.0.0
python-dateutil==2.9.0.post0
six==1.17.0
websockets==16.0
